Mark components as loaded to prevent extraneous work
The purpose of comploader is to only load components once
and to quickly execute a passed callback if the component is
already loaded.

Components were not marked as loaded, so this didn't work as
expected. However, resources (css/js files) were never loaded
more than once, because we keep track of them and prevent that
from happening.

We just don't need to pass through that no-op resource loading
code, before we execute the callback. We no longer do now.